Added parent class "Contact.java" for Address, Email, and Phone
